Fix server crash when response undefined
On occasion, the server would crash when trying to process the response to a static file.
Rather than returning it through the static middleware, it would be handled by the
housing_api route. I'm not sure which method was calling it, but the request would end up
trying to return a response that was undefined (passed as an argument to resultToResponse).

The fix was simply to check for undefined (or, if it so happens, null) responses
before attempting to write them. (An empty string is returned in their stead.)