Restore normal baseline for :after text on group name.
The issue here is that the content in the :after pseudoclass is
causing the height of the group name element to expand by a single
pixel. This is caused by the text being on the same baseline as
the group name for some reason. Moving the baseline up by one pixel
prevents the height from changing, but is noticable. To counteract
this, a cosmetic change is made by changing the position to relative
and moving the top down by one pixel. It appears as if nothing changed,
and the height remains the same.