Use TameJS+CoffeeScript to get proper startup sequence
This change ensures that startup actions happen in the proper order:

- Rooms are loaded/activated after tab is loaded
- Starred rooms "re-starred" and last active tab activated only after
  all tabs have been loaded

In implementing the proper startup sequence, we fix #4 (probabilities
are now loaded when page is reloaded to the star tab).

To get this behavior, we switch to a fork of the CoffeeScript language
that incorporates the functionality of the TameJS library (http://tamejs.org).
This fork (currently) lives at: https://github.com/maxtaco/coffee-script

It introduces the `await` keyword and `defer` function to the language.
For more information, see
https://github.com/maxtaco/coffee-script/blob/tame/TAME.md