Prevent TableSorter errors on single-column rows
When sort was triggered on rows that only had one column (these are
the placeholder messages), TableSorter tried to look for columns
beyond the first one and subsequently produced exceptions.

To address this, the rows were padded with extra, empty columns.
This is somewhat hacky, but it makes TableSorter happy without having
to change the actual library.