Fix #30 (placeholder message reappearing after deletion)
The problem was occurring on startup. One of the placeholders
("no results yet") was getting cached by the TableSorter plugin,
and the 'update' event (which told TableSorter that the content had
changed) wasn't triggered soon enough.* So when, subsequently,
sorting happened, TableSorter appended new content to old content,
resulting in both placeholders (or, in some cases, the placeholder
and result rows) being displayed.

(* - This happens, specifically, because the 'update' functionality
in TableSorter is triggered only after a timeout of 1.
Why this is the case, I don't know.)

To address this, the sorting on startup now happens with a minuscule
delay. This sequences the events such that the update gets to happen
first, so that the cache is in order when the sort happens afterwards.