Push availability changes from server to clients
... for rooms taken in the lottery.

Whenever the storage subsystem discovers than an available room
has been taken, it publishes that room to a Redis channel.

The main server is subscribed to this channel and, when it receives
a message, it pushes the room out to all the clients, which are
connected to the server using socket.io.

The clients receive this information and update the availability
of the room in question.